refactor: Move loading_messages.py to messaging module to reduce root clutter
